Available Times
Lets set up a couple days a week where we can exchange ideas, and talk
about the project. We should make it a day before the weekend so we have
enough time to work on the stuff we discuss before the next class
meeting. I think Wednesdays' and either Friday or Saturday would be
ideal. Two days a week would be good, and we should mix it up between in
person and online meetings.